Rename scan_path_storage to storage
